move geometric calculation of volume to ProductBased (prevents ciruclar imports and less redundancy)
